# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agent Planner API",
    "version": "2.0.0",
    "description": "Collaborative planning API for humans and AI agents"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
...
  },
  "paths": {
...
    "/v1/decisions": {
      "get": {
        "summary": "Pending items needing human review across accessible plans",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Pending decisions",
            "agent requests": null,
            "and drafts": null
          }
        }
      }
    },
    "/v1/decisions/{id}/resolve": {
      "post": {
        "summary": "Resolve a decision (approve/defer/reject, materialize subtasks)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Decision resolved"
          },
          "404": {
            "description": "Decision not found"
          }
        }
      }
    },
    "/v1/decisions/{id}/cancel": {
      "post": {
        "summary": "Cancel a pending decision",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Decision cancelled"
          },
          "404": {
            "description": "Decision not found"
          }
        }
      }
    },
    "/v1/dependencies": {
      "post": {
        "summary": "Create a dependency edge between two nodes (intra- or cross-plan)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Dependency created"
          },
          "409": {
            "description": "Edge would create a cycle"
          }
        }
      }
    },
    "/v1/dependencies/{id}": {
      "delete": {
        "summary": "Remove a dependency edge",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Dependency removed"
          },
          "404": {
            "description": "Dependency not found"
          }
        }
      }
    },
    "/v1/goals/dashboard": {
      "get": {
        "summary": "Goal health dashboard (on_track / at_risk / stale rollup)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goal health rollup"
          }
        }
      }
    },
    "/v1/goals": {
      "get": {
        "summary": "List goals",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goal list"
          }
        }
      },
      "post": {
        "summary": "Create a goal",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Goal created"
          }
        }
      }
    },
    "/v1/goals/{id}/state": {
      "get": {
        "summary": "Comprehensive single-goal read — details, quality, progress, bottlenecks, knowledge gaps",
        "description": "Composed facade mirroring the MCP `goal_state` tool. Partial backend failures are surfaced in meta.failures.",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Composed goal state"
          },
          "404": {
            "description": "Goal not found"
          }
        }
      }
    },
    "/v1/goals/{id}/promote": {
      "post": {
        "summary": "Promote a desire goal to an intention (readiness-checked)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Promotion result with readiness gaps if not ready"
          }
        }
      }
    },
    "/v1/goals/{id}": {
      "get": {
        "summary": "Get a goal",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goal"
          }
        }
      },
      "patch": {
        "summary": "Update a goal",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated goal"
          }
        }
      },
      "delete": {
        "summary": "Delete a goal (soft delete)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goal deleted"
          }
        }
      }
    },
...
    "/goals/dashboard": {
      "get": {
        "summary": "Get health dashboard for all user goals",
        "description": "Returns health status for all goals the authenticated user owns,\nincluding linked plan progress, bottleneck summaries, pending decisions,\nand last agent activity timestamps.\n\nHealth heuristics:\n- **stale** — no log activity on any linked plan in 3+ days\n- **at_risk** — has bottlenecks OR blocked tasks > 30% OR pending decisions older than 1 day\n- **on_track** — everything else\n",
        "tags": [
          "Goals"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Goal dashboard",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "goals": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "title": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "type": {
                            "type": "string",
                            "enum": [
                              "outcome",
                              "constraint",
                              "metric",
                              "principle"
                            ]
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "achieved",
                              "paused",
                              "abandoned"
                            ]
                          },
                          "health": {
                            "type": "string",
                            "enum": [
                              "on_track",
                              "at_risk",
                              "stale"
                            ]
                          },
                          "bottleneck_summary": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "node_id": {
                                  "type": "string"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "status": {
                                  "type": "string"
                                },
                                "direct_downstream_count": {
                                  "type": "integer"
                                }
                              }
                            }
                          },
                          "knowledge_gap_count": {
                            "type": "integer"
                          },
                          "last_activity": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "linked_plan_progress": {
                            "type": "object",
                            "properties": {
                              "total_nodes": {
                                "type": "integer"
                              },
                              "completed_nodes": {
                                "type": "integer"
                              },
                              "blocked_nodes": {
                                "type": "integer"
                              },
                              "percent_completed": {
                                "type": "number"
                              },
                              "percent_blocked": {
                                "type": "number"
                              },
                              "linked_plan_count": {
                                "type": "integer"
                              }
                            }
                          },
                          "pending_decision_count": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}